+++ craft_application/commands/lifecycle.py
@@ -28,6 +28,7 @@ from typing_extensions import override
 
 from craft_application import errors, models, util
 from craft_application.commands import base
+from craft_application.util.logging import handle_runtime_error
 
 _PACKED_FILE_LIST_PATH = ".craft/packed-files"
 
@@ -161,13 +162,43 @@ class LifecycleCommand(_BaseLifecycleCommand):
                 metavar="arch",
                 help="Set architecture to build for",
             )
+        ignore_restrictions: set[str] = set()
         if self._app.check_supported_base:
+            ignore_restrictions.add("unmaintained")
+        if ignore_restrictions:
             parser.add_argument(
-                "--old-bases",
-                action="store_true",
-                dest="allow_unsupported_base",
-                help="Allow this run to work on an unsupported base.",
+                "--ignore",
+                nargs="*",
+                default=[],
+                choices=sorted(ignore_restrictions),
+                help="Bypass a restriction to use an unsupported feature.",
             )
+
+    def _check_supported_base(self, parsed_args: argparse.Namespace) -> None:
+        """Check whether the base is supported.
+
+        :raises: CraftValidationError if building on an unsupported base without
+            explicitly acknowledging that fact.
+        """
+        if not self._app.check_supported_base:
+            return
+        project_service = self._services.get("project")
+        project = project_service.get()
+        ignore_items = getattr(parsed_args, "ignore", [])
+        if "unmaintained" in ignore_items:
+            if project_service.is_effective_base_eol():
+                emit.progress(
+                    f"Using end-of-life base {project.effective_base!r}.",
+                    permanent=True,
+                )
+        else:
+            project_service.check_base_is_supported(verb=self.name)
+            base = project.effective_base
+            if eol_date := project_service.base_eol_soon_date():
+                emit.progress(
+                    f"Using base {base!r}, which will enter end-of-life on {eol_date.isoformat()}.",
+                    permanent=True,
+                )
 
     @override
     def _run(
@@ -182,10 +213,7 @@ class LifecycleCommand(_BaseLifecycleCommand):
         build_planner = self.services.get("build_plan")
 
         # If we check for supported bases, fail early for all unsupported bases.
-        if self._app.check_supported_base and not getattr(
-            parsed_args, "allow_unsupported_base", None
-        ):
-            self.services.get("project").check_base_is_supported()
+        self._check_supported_base(parsed_args)
 
         config = self.services.get("config")
         platform = getattr(parsed_args, "platform", None) or config.get("platform")
@@ -221,7 +249,7 @@ class LifecycleCommand(_BaseLifecycleCommand):
             self._run_lifecycle(parsed_args, step_name)
         except Exception as err:
             if debug:
-                emit.progress(str(err), permanent=True)
+                handle_runtime_error(self._app, err)
                 _launch_shell()
             raise
 
@@ -654,7 +682,7 @@ class TestCommand(PackCommand):
         provider = self._services.get("provider")
 
         fetch_service_policy = cast(
-            Literal["strict", "permissive", None],
+            Literal["strict", "permissive"] | None,
             getattr(parsed_args, "fetch_service_policy", None),
         )
         if fetch_service_policy: