+++ witchcraft/application.py
@@ -19,15 +19,20 @@ import craft_application
 from craft_application.application import Application
 from typing_extensions import override
 
+from .models import Project
+
 WITCHCRAFT = craft_application.AppMetadata(
     name="witchcraft",
     summary="A craft for testing craft-application with weird settings",
+    artifact_type="cauldron",
     docs_url="https://canonical-craft-application.readthedocs-hosted.com",
     source_ignore_patterns=["*.witchcraft", "witchcraft.yaml"],
     project_variables=["version"],
     mandatory_adoptable_fields=["version"],
     always_repack=False,
     check_supported_base=True,
+    enable_for_grammar=True,
+    ProjectClass=Project,
 )